Update LocationUtil : Refactored Logic with Koin AndroidContext Injection for Android Module
